requirements.txt
Django==1.9.2
djangorestframework
markdown
chardet
requests
psycopg2
pyyaml
pytz
mixer
pycrypto
celery
django-celery
flower
